UT on ServeUnique
