+++ htagweb/__init__.py
@@ -7,9 +7,9 @@
 # https://github.com/manatlan/htagweb
 # #############################################################################
 
-
 from .appserver import AppServer  # a completly different beast, but compatible with ^^
 from .htagserver import HtagServer  # a completly different beast.
+from .serveunique import ServeUnique
 
 __all__= ["AppServer"]
